+++ bin/systemInfo.js
@@ -46,18 +46,23 @@ exports.handler = function (argv) {
 
 
 async function dbStatus(result) {
+  try {
+    const dbStatus = new dbClass(await dbClass.createConnectionFromEnv(dbClass.resolveEnv(result)));
+    const dbInspect = require("../utils/dbInspect")
+    console.table(await dbInspect.getHANAVersion(dbStatus))
 
-  const dbStatus = new dbClass(await dbClass.createConnectionFromEnv(dbClass.resolveEnv(result)));
-  const dbInspect = require("../utils/dbInspect")
-  console.table(await dbInspect.getHANAVersion(dbStatus))
+    let results = await dbStatus.execSQL(`SELECT TOP 100 * FROM "M_SYSTEM_OVERVIEW"`)
 
-  let results = await dbStatus.execSQL(`SELECT TOP 100 * FROM "M_SYSTEM_OVERVIEW"`);
+    console.table(results)
 
-  console.table(results);
+    results = await dbStatus.execSQL(`SELECT TOP 100 * FROM "M_SERVICES"`)
+    console.table(results)
 
-  results = await dbStatus.execSQL(`SELECT TOP 100 * FROM "M_SERVICES"`);
-  console.table(results);
+    global.__spinner.stop()
+    return
+  } catch (error) {
+    global.__spinner.stop()
+    console.error(`Connection Problem ${error}`)
 
-  global.__spinner.stop()
-  return;
+  }
 }